requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-telegram-bot==21.11.1
httpx==0.28.1
pydantic>=2.12.0
python-dotenv==1.1.0
pymysql>=1.1.0
